requirements.txt
Django==1.10.5
django-cors-headers==2.0.1
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-filter==1.0.1
django-ipware==1.1.6
django-model-utils==2.6.1
django-rest-swagger==2.1.1
djangorestframework-jwt==1.9.0
djangorestframework==3.5.3
gunicorn==19.6.0
json-logging-py==0.2
psycopg2==2.6.2
pytz==2016.10
whitenoise==3.3.0
